Use 64 bits to compute the number of voxels while reading file

+++ yrt-pet/src/datastruct/image/Image.cpp
@@ -1156,73 +1156,74 @@ void ImageOwned::readNIfTIData(int datatype, void* data, float slope,
 	const ImageParams& params = getParams();
 
 	float* imgData = getRawPointer();
-	const int numVoxels = getNumFrames() * params.nx * params.ny * params.nz;
+	const ssize_t numVoxels =
+	    getNumFrames() * params.nx * params.ny * params.nz;
 
 	if (datatype == NIFTI_TYPE_FLOAT32)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (*(reinterpret_cast<float*>(data) + i) * slope) + intercept;
 	}
 	else if (datatype == NIFTI_TYPE_FLOAT64)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<double, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_INT8)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<int8_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_INT16)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<int16_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_INT32)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<int32_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_INT64)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<int64_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_UINT8)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<uint8_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_UINT16)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<uint16_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_UINT32)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<uint32_t, float>(data, i) * slope) +
 			    intercept;
 	}
 	else if (datatype == NIFTI_TYPE_UINT64)
 	{
-		for (int i = 0; i < numVoxels; i++)
+		for (ssize_t i = 0; i < numVoxels; i++)
 			imgData[i] =
 			    (util::reinterpretAndCast<uint64_t, float>(data, i) * slope) +
 			    intercept;
